refactor(tracker): reorganize ml-fetch into multiple files
