+++ modules/tracker/apps/ml-fetch/src/index.ts
@@ -3,27 +3,22 @@
 import { rawVehicleEventsNew } from '@tmlmobilidade/databases';
 import { Dates } from '@tmlmobilidade/dates';
 import { externalClients } from '@tmlmobilidade/external';
-import { BaseResponse, DESTINATION_MAP, TempoEsperaRawItem } from '@tmlmobilidade/external/dist/clients/ml/types.js';
-import { chunkLineByDistanceV2, hashedShapesToFeatureCollection, lineSlice, nearestPointOnLine, point } from '@tmlmobilidade/geo';
-import { rides, stops } from '@tmlmobilidade/interfaces';
+import { type BaseResponse, type TempoEsperaRawItem } from '@tmlmobilidade/external/dist/clients/ml/types.js';
 import { initSentryNode, Logger } from '@tmlmobilidade/logger';
 import { Timer } from '@tmlmobilidade/timer';
 import { type HashableRawVehicleEvent, type RawVehicleEventMlV1 } from '@tmlmobilidade/types';
 import { runOnInterval } from '@tmlmobilidade/utils';
-import { Feature, LineString, Point } from 'geojson';
 import crypto from 'node:crypto';
 
-import { aggregationQuery } from './aggregation-query.js';
-import { AggregationResult, TripStopWaypoint } from './types.js';
+import { findRideForTrain } from './find-ride-for-train.js';
+import { findTripStopWaypoints } from './find-trip-stop-waypoints.js';
+import { groupTrainPositions } from './group-train-positions.js';
+import { inferTrainPositionOnShape } from './infer-train-position.js';
+import { AggregationResult, type TrainPositionsMap } from './types.js';
 
 /* * */
 
 let ITERATION = 0;
-
-const gtfsTimeToSeconds = (time: string): number => {
-	const [hours, minutes, seconds] = time.split(':').map(Number);
-	return hours * 3600 + minutes * 60 + seconds;
-};
 
 /* * */
 
@@ -50,7 +45,7 @@ const main = async () => {
 	// Initialize the train positions map.
 	// This groups the upcoming trains per line so we can infer where each one is.
 
-	const trainPositionsMap = new Map<string, { destination_id: string, next_stop: { arrival_seconds: number, stop_id: string } }>();
+	const trainPositionsMap: TrainPositionsMap = new Map();
 
 	//
 	// Fetch the Metro Lisboa Vehicle Events data from API and decode it.
@@ -79,38 +74,7 @@ const main = async () => {
 			continue;
 		}
 
-		//
-		// Group the response by train.
-		// Each platform item reports up to 3 upcoming trains with their arrival times.
-
-		for (const item of response.resposta) {
-			//
-
-			//
-			// Skip invalid data.
-
-			if (!item.comboio || item.tempoChegada1 === '--') continue;
-
-			//
-			// Parse the data.
-
-			const trainId = item.comboio;
-			const destinationId = DESTINATION_MAP[item.destino as unknown as keyof typeof DESTINATION_MAP]?.code;
-			const arrivalSeconds = Number.parseInt(item.tempoChegada1);
-			const stopId = item.stop_id;
-
-			//
-			// If there is no entry for this train, create one.
-			if (!trainPositionsMap.has(trainId)) {
-				trainPositionsMap.set(trainId, { destination_id: destinationId, next_stop: { arrival_seconds: arrivalSeconds, stop_id: stopId } });
-			}
-
-			//
-			// If the new arrival time is earlier than the current one, update the entry, as we only care about the earliest arrival time.
-			if (arrivalSeconds < trainPositionsMap.get(trainId).next_stop.arrival_seconds) {
-				trainPositionsMap.get(trainId).next_stop = { arrival_seconds: arrivalSeconds, stop_id: stopId };
-			}
-		}
+		groupTrainPositions({ items: response.resposta, trainPositionsMap });
 
 		//
 		// For each train, infer its current position.
@@ -118,133 +82,36 @@ const main = async () => {
 		for (const [trainId, { destination_id: destinationId, next_stop: nextStop }] of trainPositionsMap) {
 			//
 
-			//
-			// Finds a single HashedTrip for this train with the destionation stop name as the trip headsign.
-			// The trip is enriched with the stop codes for each waypoint in the path.
-			// This will be used to understand the time difference between the previous and the trains `next_stop`.
-
+			/**
+			 * For each train, use findRideForTrain to match a Metro Lisboa ride based on the inferred destination (from destinationId)
+			 * and the current reference time (now). If no corresponding ride is found for the current train, skip to the next train.
+			 * Any error while searching is logged, and the loop continues.
+			 */
 			let ride: AggregationResult | null = null;
 			try {
-				// Find the GTFS stop document for the train's destination
-				const destinationStop = await stops.findOne({
-					flags: { $elemMatch: { agency_ids: '2', stop_id: destinationId } },
-				});
-
-				if (!destinationStop) continue;
-
-				// Fetch hashed trip patterns matching this trip's headsign and arrival window
-				const ridesCollection = await rides.getCollection();
-
-				const ridesAggregationResult = await ridesCollection.aggregate(
-					aggregationQuery({
-						endTimeScheduled: now.plus({ hours: 2 }).unix_timestamp,
-						headsign: destinationStop.name,
-						startTimeScheduled: now.minus({ hours: 2 }).unix_timestamp,
-					}),
-				).toArray() as AggregationResult[];
-
-				if (ridesAggregationResult.length === 0) continue;
-				// Use the first match (should only be one due to unique pattern).
-				ride = ridesAggregationResult[0];
+				ride = await findRideForTrain({ destinationId, now });
 			} catch (error) {
-				Logger.error({ error, message: `[${ITERATION}] Error finding hashed trip for train ${trainId} on line ${line}:` });
-				console.log(error);
+				Logger.error({ error, message: `[${ITERATION}] Error finding ride for train ${trainId} on line ${line}:` });
 				continue;
 			}
 
-			//
-			// Find the next stop's waypoint in the hashed trip path.
+			console.log(JSON.stringify(ride, null, 2));
 
-			let previousStopWaypoint: TripStopWaypoint | undefined = undefined;
-			let nextStopWaypoint: TripStopWaypoint | undefined = undefined;
-			for (const [index, waypoint] of ride.hashed_trip.path.entries()) {
-				const foundStopCode = waypoint.stop_codes.find(code => code === nextStop.stop_id);
-				if (!foundStopCode) continue;
+			if (!ride) continue;
 
-				if (index > 0) {
-					const timeDifference = gtfsTimeToSeconds(waypoint.arrival_time) - gtfsTimeToSeconds(ride.hashed_trip.path[index - 1].departure_time);
+			// Given the next stop (from the API) and the matched ride document, extract the waypoints in the ride's shape that correspond to:
+			// - nextStopWaypoint: The shape point nearest to the nextStop.
+			// - previousStopWaypoint: The shape point immediately preceding the nextStop (used to interpolate train position).
 
-					previousStopWaypoint = {
-						latitude: ride.hashed_trip.path[index - 1].stop_lat,
-						longitude: ride.hashed_trip.path[index - 1].stop_lon,
-						stop_id: ride.hashed_trip.path[index - 1].stop_id,
-						timeDifference,
-					};
+			const { nextStopWaypoint, previousStopWaypoint } = findTripStopWaypoints({ nextStop, ride });
 
-					nextStopWaypoint = {
-						latitude: waypoint.stop_lat,
-						longitude: waypoint.stop_lon,
-						stop_id: waypoint.stop_id,
-						timeDifference,
-					};
-				} else {
-					nextStopWaypoint = {
-						latitude: waypoint.stop_lat,
-						longitude: waypoint.stop_lon,
-						stop_id: waypoint.stop_id,
-						timeDifference: 0,
-					};
-				}
-
-				break;
-			}
+			console.log(JSON.stringify({ nextStopWaypoint, previousStopWaypoint }, null, 2));
 
 			if (!nextStopWaypoint) continue;
 
-			//
-			// Infer position along the shape. When the next stop is the first in the trip,
-			// there is no previous stop to interpolate from — place the train at that stop.
-
-			let pointInChunkedLine: [number, number];
-
-			if (!previousStopWaypoint) {
-				pointInChunkedLine = [nextStopWaypoint.longitude, nextStopWaypoint.latitude];
-			} else {
-				//
-				// Convert HashedShape to GeoJSON
-				const hashedShapeLineString: Feature<LineString> = hashedShapesToFeatureCollection(ride.hashed_shape).features[0];
-				const previousStopPoint: Feature<Point> = point([previousStopWaypoint.longitude, previousStopWaypoint.latitude]);
-				const nextStopPoint: Feature<Point> = point([nextStopWaypoint.longitude, nextStopWaypoint.latitude]);
-
-				const nearestPreviousStopPointOnShape = nearestPointOnLine(hashedShapeLineString, previousStopPoint);
-				const nearestNextStopPointOnShape = nearestPointOnLine(hashedShapeLineString, nextStopPoint);
-
-				// Subsection of shape between the two nearest points
-				const slicedShape = lineSlice(nearestPreviousStopPointOnShape, nearestNextStopPointOnShape, hashedShapeLineString);
-				const chuckedLine = chunkLineByDistanceV2(slicedShape.geometry, 25);
-
-				if (chuckedLine.coordinates.length === 0 || nextStopWaypoint.timeDifference <= 0) {
-					pointInChunkedLine = [
-						(previousStopWaypoint.longitude + nextStopWaypoint.longitude) / 2,
-						(previousStopWaypoint.latitude + nextStopWaypoint.latitude) / 2,
-					];
-				} else {
-					// Given the time difference between the previous and the next stop,
-					// And the next_stop: arrival_seconds we can calculate the percentage of the chunked line that the vehicle has traveled.
-					// Eg. If ETA is 30s (nextStop.arrival_seconds) and delta between stops (nextStopWaypoint.timeDifference) is 120s
-					// This means the vehicle has traveled 75% of the distance between the two stops, and it's missing the last 25%.
-					// ((nextStopWaypoint.timeDifference - nextStop.arrival_seconds))/nextStopWaypoint.timeDifference = percentage of the chunked line that the vehicle has traveled.
-
-					const timeDifferenceBetweenPreviousAndNextStop = nextStopWaypoint.timeDifference - nextStop.arrival_seconds;
-					let percentageOfChunkedLine = timeDifferenceBetweenPreviousAndNextStop / nextStopWaypoint.timeDifference;
-
-					//
-					// If for some reason the percentage is lower than 0%, means that the vehicle is delayed
-					// Here we'll set the percentage to 0.5, which means that the vehicle is halfway between the two stops.
-					// (We're not sure if this is correct, but it's a guess)
-
-					if (percentageOfChunkedLine < 0) percentageOfChunkedLine = 0.5;
-
-					const coordinateIndex = Math.min(
-						Math.max(0, Math.floor(percentageOfChunkedLine * (chuckedLine.coordinates.length - 1))),
-						chuckedLine.coordinates.length - 1,
-					);
-
-					pointInChunkedLine = chuckedLine.coordinates[coordinateIndex] as [number, number];
-				}
-			}
-
-			if (!pointInChunkedLine) continue;
+			// Infer the train's current position as a point [longitude, latitude] on the ride's shape,
+			// using the next stop, corresponding waypoints, and ride details.
+			const pointInChunkedLine = inferTrainPositionOnShape({ nextStop, nextStopWaypoint, previousStopWaypoint, ride });
 
 			//
 			// Hash the relevant fields of the vehicle event
